Refactor TransactionActions to have explicitly named single cell variants
