+++ atlasdb-workload-server-api/src/main/java/com/palantir/atlasdb/workload/transaction/TransactionAction.java
@@ -16,18 +16,6 @@
 
 package com.palantir.atlasdb.workload.transaction;
 
-import com.palantir.atlasdb.workload.store.WorkloadCell;
-import org.immutables.value.Value;
-
 public interface TransactionAction {
-
-    /** Table to apply transaction to. */
-    @Value.Parameter
-    String table();
-
-    /** Cell (Key, Column) to apply the action to. */
-    @Value.Parameter
-    WorkloadCell cell();
-
     <T> T accept(TransactionActionVisitor<T> visitor);
 }
