+++ atlasdb-workload-server-api/src/main/java/com/palantir/atlasdb/workload/transaction/witnessed/WitnessedWriteTransactionAction.java
@@ -20,7 +20,7 @@ import com.palantir.atlasdb.workload.store.WorkloadCell;
 import org.immutables.value.Value;
 
 @Value.Immutable
-public interface WitnessedWriteTransactionAction extends WitnessedTransactionAction {
+public interface WitnessedWriteTransactionAction extends WitnessedSingleCellTransactionAction {
 
     @Override
     @Value.Parameter
